requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.17.1
python-docx>=1.1.0
openai>=1.3.7
streamlit>=1.28.1
pandas>=2.1.3
numpy>=1.26.0
docx2txt>=0.8
jieba>=0.42.1
scikit-learn>=1.3.0 